Update room header styles (#13)
