+++ src/pages/Room/_Header.js
@@ -1,27 +1,21 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import history from '../../app/history'
-import { leaveRoomChannel } from '../../actions/rooms'
-import { deleteSubscription } from '../../actions/userSubscriptions'
 import { getCurrentUser } from '../../reducers/currentUser'
 import { getDirectMessageUser } from '../../reducers/roomSubscriptions'
 import { getIsSubscribed, getRoom } from '../../reducers/userSubscriptions'
-import notification from '../../helpers/notification'
-import { rootPath } from '../../helpers/paths'
 import Header from '../Layout/_Header'
+import RoomHeaderActions from './_HeaderActions'
 
-const RoomHeader = ({ handleLeave, heading, isDirectMessage, isSubscribed }) => {
-  return (
-    <Header>
-      <h2>{ heading }</h2>
-      <div className='chat-room-header-actions'>
-        { isSubscribed && !isDirectMessage &&
-          <button onClick={ handleLeave }>Leave</button>
-        }
-      </div>
-    </Header>
-  )
-}
+const RoomHeader = ({ heading, isDirectMessage, isSubscribed, room }) => (
+  <Header>
+    <h2>{ heading }</h2>
+    <RoomHeaderActions
+      room={ room }
+      isSubscribed={ isSubscribed }
+      isDirectMessage={ isDirectMessage }
+    />
+  </Header>
+)
 
 RoomHeader.displayName = 'RoomHeader'
 
@@ -38,20 +32,8 @@ const mapStateToProps = (state, { room: slug }) => {
   }
 }
 
-const mapDispatchToProps = (dispatch, { room: slug }) => ({
-  handleLeave: () => {
-    const onSuccess = () => {
-      dispatch(leaveRoomChannel(slug))
-      notification('Left room ' + slug, 'success')
-      history.push(rootPath)
-    }
+const mapDispatchToProps = () => ({
 
-    const onError = () => {
-      notification('Error leaving room ' + slug, 'error')
-    }
-
-    return dispatch(deleteSubscription(slug, onSuccess, onError))
-  }
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(RoomHeader)
